+++ src/NetworkAdapter.test.ts
@@ -1,6 +1,11 @@
 import { describe, expect, it } from "vitest";
 
 describe("NetworkAdapter", () => {
+  /**
+   * NOTE: WebRTC is a browser based technology so server-side testing is hard.
+   *       Over time, if it becomes necessary, we can expand these tests out
+   *       to include mocking.
+   */
   it("placeholder", () => {
     expect(123).to.eql(123);
   });