+++ src/NetworkAdapter.ts
@@ -1,6 +1,7 @@
-import { NetworkAdapter } from "@automerge/automerge-repo";
 import { EventEmitter } from "eventemitter3";
-import type * as t from "./t.js";
+import type * as t from "./types.js";
+
+type EventTypes = { data: t.NetworkMessageAlert };
 
 /**
  * An Automerge repo network-adapter for WebRTC (P2P)
@@ -10,10 +11,16 @@ import type * as t from "./t.js";
  *    https://github.com/automerge/automerge-repo/blob/main/packages/automerge-repo-network-messagechannel/src/index.ts
  *
  */
-export class WebrtcNetworkAdapter extends NetworkAdapter {
-  #conn: t.DataConnection;
+export class PeerjsNetworkAdapter
+  extends EventEmitter<t.NetworkAdapterEvents>
+  implements t.NetworkAdapterInterface
+{
+  peerId?: t.PeerId;
+  peerMetadata?: t.PeerMetadata;
+
   #isReady = false;
-  #disconnected = new EventEmitter<"disconnected">();
+  #conn: t.DataConnection;
+  #events = new EventEmitter<EventTypes>();
 
   constructor(conn: t.DataConnection) {
     if (!conn) throw new Error(`A peerjs data-connection is required`);
@@ -21,14 +28,15 @@ export class WebrtcNetworkAdapter extends NetworkAdapter {
     this.#conn = conn;
   }
 
-  connect(peerId: t.PeerId) {
+  connect(peerId: t.PeerId, meta?: t.PeerMetadata) {
     const senderId = (this.peerId = peerId);
     const conn = this.#conn;
+    const peerMetadata = meta ?? {};
 
-    const handleOpen = () => this.#transmit({ type: "arrive", senderId, peerMetadata: {} });
+    const handleOpen = () => this.#transmit({ type: "arrive", senderId, peerMetadata });
     const handleClose = () => this.emit("close");
     const handleData = (e: any) => {
-      const msg = e as t.WebrtcMessage;
+      const msg = e as t.NetworkMessage;
 
       /**
        * Arrive.
@@ -56,13 +64,14 @@ export class WebrtcNetworkAdapter extends NetworkAdapter {
       let payload = msg as t.Message;
       if ("data" in msg) payload = { ...payload, data: toUint8Array(msg.data!) };
       this.emit("message", payload);
+      this.#alert("incoming", msg);
     };
 
     conn.on("open", handleOpen);
     conn.on("close", handleClose);
     conn.on("data", handleData);
 
-    this.#disconnected.on("disconnected", () => {
+    this.on("peer-disconnected", () => {
       this.#isReady = false;
       conn.off("open", handleOpen);
       conn.off("close", handleClose);
@@ -78,7 +87,13 @@ export class WebrtcNetworkAdapter extends NetworkAdapter {
   }
 
   disconnect() {
-    this.#disconnected.emit("disconnected");
+    const peerId = this.peerId;
+    if (peerId) this.emit("peer-disconnected", { peerId });
+  }
+
+  onData(fn: (e: t.NetworkMessageAlert) => void) {
+    this.#events.on("data", fn);
+    return () => this.#events.off("data", fn);
   }
 
   send(message: t.RepoMessage) {
@@ -90,9 +105,16 @@ export class WebrtcNetworkAdapter extends NetworkAdapter {
     }
   }
 
-  #transmit(message: t.WebrtcMessage) {
+  #transmit(message: t.NetworkMessage) {
     if (!this.#conn) throw new Error("Connection not ready");
     this.#conn.send(message);
+    this.#alert("outgoing", message);
+  }
+
+  #alert(direction: t.IODirection, message: t.NetworkMessage) {
+    const bytes = "data" in message ? message.data?.byteLength ?? 0 : 0;
+    const payload: t.NetworkMessageAlert = { direction, message, bytes };
+    this.#events.emit("data", payload);
   }
 
   #setAsReady() {