+++ src/types.ts
@@ -1,8 +1,22 @@
 /**
  * @automerge
  */
-import type { Message, PeerId, RepoMessage, StorageId } from "@automerge/automerge-repo";
-export { Message, PeerId, RepoMessage };
+import type {
+  Message,
+  NetworkAdapterEvents,
+  NetworkAdapterInterface,
+  PeerId,
+  RepoMessage,
+  StorageId,
+} from "@automerge/automerge-repo";
+export type {
+  Message,
+  NetworkAdapterEvents,
+  NetworkAdapterInterface,
+  PeerId,
+  RepoMessage,
+  StorageId,
+};
 
 /**
  * @peerjs
@@ -15,12 +29,14 @@ export type { DataConnection } from "peerjs";
  *    MessageChannelNetworkAdapter
  *    https://github.com/automerge/automerge-repo/blob/main/packages/automerge-repo-network-messagechannel/src/index.ts
  */
-export type IODirection = "incoming" | "outgoing";
+export type { PeerjsNetworkAdapter } from "./NetworkAdapter.js";
 
-export type WebrtcMessage = ArriveMessage | WelcomeMessage | Message;
-export type WebrtcMessageAlert = {
+export type IODirection = "incoming" | "outgoing";
+export type NetworkMessage = ArriveMessage | WelcomeMessage | Message;
+export type NetworkMessageAlert = {
   direction: IODirection;
-  message: WebrtcMessage;
+  message: NetworkMessage;
+  bytes: number;
 };
 
 /**